backend_repo/requirements.txt
Django>=5.1
djangorestframework>=3.14
